Updated off/online.sh
Only allow offline to create one backup; only revert and reboot if offline.sh has been called.

+++ src/scripts/misc/offline.sh
@@ -1,4 +1,8 @@
-mv /home/tesla/.Tesla/car/cell_apn /home/tesla/.Tesla/car/cell_apn.bak
-touch /home/tesla/.Tesla/car/cell_apn
-echo "junk" >> /home/tesla/.Tesla/car/cell_apn
-emit-reboot-gateway
+#!/bin/bash
+
+# Do not overwrite back up if already taken.
+if [ ! -f /home/tesla/.Tesla/car/cell_apn.bak ]; then
+	mv /home/tesla/.Tesla/car/cell_apn /home/tesla/.Tesla/car/cell_apn.bak
+	echo "junk" > /home/tesla/.Tesla/car/cell_apn
+	emit-reboot-gateway
+fi
